move Symfony rector to Contrib
