+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/Source/LocalKernel.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\NodeVisitor\DependencyInjection\NamedServicesToConstructorRector\Source;
+namespace Rector\Tests\Rector\Contrib\Symfony\GetterToPropertyRector\Source;
 
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
