+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/Source/services.yml
