+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/Test.php
@@ -1,7 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\NodeVisitor\DependencyInjection\NamedServicesToConstructorRector;
+namespace Rector\Tests\Rector\Contrib\Symfony\GetterToPropertyRector;
 
+use Rector\Rector\Contrib\Symfony\GetterToPropertyRector;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
 final class Test extends AbstractRectorTestCase
@@ -29,6 +30,6 @@ final class Test extends AbstractRectorTestCase
      */
     protected function getRectorClasses(): array
     {
-        return [];
+        return [GetterToPropertyRector::class];
     }
 }
