+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/correct/correct.php.inc
