+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/correct/correct2.php.inc
