+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/correct/correct3.php.inc
