+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/wrong/wrong.php.inc
