+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/wrong/wrong2.php.inc
