+++ tests/Rector/Contrib/Symfony/GetterToPropertyRector/wrong/wrong3.php.inc
